+++ lib/karafka/railtie.rb
@@ -1,9 +1,30 @@
 # frozen_string_literal: true
 
 # This file contains Railtie for auto-configuration
+
+rails = false
+
 begin
+  # We check this instead of rails as that way we can support piecemeal rails setup
+  require 'railties'
+
+  rails = true
+rescue LoadError
+  # Without defining this in any way, Zeitwerk ain't happy so we do it that way
+  module Karafka
+    class Railtie
+    end
+  end
+end
+
+if rails
+  # Load Karafka
   require 'karafka'
-  require 'rails'
+  # Load ActiveJob adapter
+  require 'active_job/karafka'
+
+  # Setup env if configured (may be configured later by .net, etc)
+  ENV['KARAFKA_ENV'] ||= ENV['RAILS_ENV'] if ENV.key?('RAILS_ENV')
 
   module Karafka
     # Railtie for setting up Rails integration
@@ -11,8 +32,10 @@ begin
       railtie_name :karafka
 
       initializer 'karafka.configure_rails_initialization' do |app|
+        # Consumers should autoload by default in the Rails app so they are visible
         app.config.autoload_paths += %w[app/consumers]
 
+        # Make Karafka use Rails logger
         ::Karafka::App.config.logger = Rails.logger
 
         # This lines will make Karafka print to stdout like puma or unicorn
@@ -25,15 +48,10 @@ begin
         end
 
         app.reloader.to_prepare do
-          require Rails.root.join(Karafka.boot_file)
+          # Load Karafka bot file, so it can be used in Rails server context
+          require Rails.root.join(Karafka.boot_file.to_s).to_s
         end
       end
-    end
-  end
-rescue LoadError
-  # Without defining this in any way, Zeitwerk ain't happy so we do it that way
-  module Karafka
-    class Railtie
     end
   end
 end