+++ spec/lib/karafka/process_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe Karafka::Process do
+RSpec.describe_current do
   subject(:process) { described_class.new }
 
   before { process.instance_variable_set(:@callbacks, Hash.new { |hsh, key| hsh[key] = [] }) }